+++ assistant/src/__tests__/messaging-skill-split.test.ts
@@ -37,8 +37,6 @@ describe("Messaging skill split", () => {
 
   const expectedGmailToolNames = [
     "gmail_archive",
-    "gmail_batch_archive",
-    "gmail_archive_by_query",
     "gmail_label",
     "gmail_batch_label",
     "gmail_trash",
