+++ assistant/src/__tests__/registry.test.ts
@@ -221,7 +221,6 @@ describe("baseline characterization: hardcoded tool loading", () => {
       "gmail_mark_read",
       "gmail_draft",
       "gmail_archive",
-      "gmail_batch_archive",
       "gmail_label",
       "gmail_batch_label",
       "gmail_trash",
@@ -241,7 +240,6 @@ describe("baseline characterization: hardcoded tool loading", () => {
       "gmail_mark_read",
       "gmail_draft",
       "gmail_archive",
-      "gmail_batch_archive",
       "gmail_label",
       "gmail_batch_label",
       "gmail_trash",
