+++ assistant/src/config/bundled-tool-registry.ts
@@ -67,8 +67,6 @@ import * as followupList from "./bundled-skills/followups/tools/followup-list.js
 import * as followupResolve from "./bundled-skills/followups/tools/followup-resolve.js";
 // ── gmail ────────────────────────────────────────────────────────────────────
 import * as gmailArchive from "./bundled-skills/gmail/tools/gmail-archive.js";
-import * as gmailArchiveByQuery from "./bundled-skills/gmail/tools/gmail-archive-by-query.js";
-import * as gmailBatchArchive from "./bundled-skills/gmail/tools/gmail-batch-archive.js";
 import * as gmailBatchLabel from "./bundled-skills/gmail/tools/gmail-batch-label.js";
 import * as gmailDownloadAttachment from "./bundled-skills/gmail/tools/gmail-download-attachment.js";
 import * as gmailDraft from "./bundled-skills/gmail/tools/gmail-draft.js";
@@ -283,8 +281,6 @@ export const bundledToolRegistry = new Map<string, SkillToolScript>([
 
   // gmail
   ["gmail:tools/gmail-archive.ts", gmailArchive],
-  ["gmail:tools/gmail-batch-archive.ts", gmailBatchArchive],
-  ["gmail:tools/gmail-archive-by-query.ts", gmailArchiveByQuery],
   ["gmail:tools/gmail-label.ts", gmailLabel],
   ["gmail:tools/gmail-batch-label.ts", gmailBatchLabel],
   ["gmail:tools/gmail-trash.ts", gmailTrash],
